refactor: document API path
