+++ src/api/document/documentRouter.ts
@@ -2,19 +2,19 @@ import type { FastifyPluginAsync } from 'fastify/types/plugin';
 
 import { createRouteSchema } from '@/libraries/httpHandlers';
 import resetDelete from '@/api/document/reset.delete';
-import parentQueryPost from '@/api/document/parent/query/post';
+import queryPost from '@/api/document/query/post';
 import { PostQuery } from '@/api/document/documentSchema';
-import parentLoadDirectoryPut from '@/api/document/parent/load/directory.put';
-import parentLoadConfluencePut from '@/api/document/parent/load/confluence.put';
+import loadDirectoryPut from '@/api/document/load/directory.put';
+import loadConfluencePut from '@/api/document/load/confluence.put';
 
 const documentRouter: FastifyPluginAsync = async fastify => {
   fastify.delete('/reset', createRouteSchema({}), resetDelete());
 
-  fastify.put('/parent/load/directory', createRouteSchema({}), parentLoadDirectoryPut());
+  fastify.put('/load/directory', createRouteSchema({}), loadDirectoryPut());
 
-  fastify.put('/parent/load/confluence', createRouteSchema({}), parentLoadConfluencePut());
+  fastify.put('/load/confluence', createRouteSchema({}), loadConfluencePut());
 
-  fastify.post('/parent/query', createRouteSchema({ body: PostQuery }), parentQueryPost());
+  fastify.post('/query', createRouteSchema({ body: PostQuery }), queryPost());
 };
 
 export default documentRouter;
